[export] Introduce build flags for shared bitgen
This cl introduces two flags to switch between two different SharedBitGen implementations.

The flags are:
* `--define grpc_bitgen_implementation=cpu_intensive` to select a CPU intensive bitgen implementation.
* `--define grpc_bitgen_implementation=mem_intensive` to select a memory intensive bitgen implementation.

If none of the two flags is provided, we'll default to the current memory intensive bitgen implementation, reserving the right to change this default in the future.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/775697153](http://cl/775697153)

PiperOrigin-RevId: 775697153