[export] Fix: Capture shared_ptr by value in channelz callback.
The lambda passed to `RunZTrace` should capture the `state` shared_ptr by value to ensure the lifetime of the state object extends until the callback is executed. Capturing by reference could lead to use-after-free if the original shared_ptr goes out of scope.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/805137760](http://cl/805137760)

PiperOrigin-RevId: 805137760