[export] Add fine-grained outstanding byte tracking for Chaotic Good Endpoints
The current implementation only collects the total amount of unsent bytes on the TCP socket every 100ms and adds to it any bytes scheduled on the endpoint between collection intervals.

In this CL, we split up this total bytes into three in order to get a finer grained picture of the data queued onto the endpoint. This will enable us to make better scheduling decisions. The three new fields are:
- `network_outstanding_bytes_`: Tracks the unsent data in the TCP socket, updated every 100ms.
- `reader_outstanding_bytes_`: Tracks the bytes scheduled on the `frames_` vector of the Data Endpoint's reader.
- `endpoint_outstanding_bytes_`: Tracks the bytes being written to the TCP socket (via a sendmsg call downstream of PromiseEndpoint::Write).

The old `last_send_outstanding_bytes_` is roughly equal to the sum of these three fields (but not exactly equal since it is reset every 100ms; there was a possibility of undercounting).

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/805655603](http://cl/805655603)

PiperOrigin-RevId: 805655603