[export] Add `ls` command to `channelz_tool` for listing entities.
This change introduces a new `ls` command to sleuth which allows listing channelz entities. The command can optionally filter by entity kind. A new flag `--channelz_columns` is added to customize the columns displayed in the output, allowing users to specify which entity properties to show. The underlying `EntityTableColumn` struct is updated to use `std::string` for better lifetime management.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/807246238](http://cl/807246238)

PiperOrigin-RevId: 807246238