[export] [Part 1] Resource Quota Write memory tracking.
Pass a MemoryAllocator to the serialize function, so that the write memory can be allocated towards Resource Quota accounting. Add templates for SerializationTraits, to allow implementations to continue using an implementation of Serialize, which does not take the allocator as a parameter. This change is a no-op for now, because all the callers of Serialize pass nullptr for the allocator.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/809691195](http://cl/809691195)

PiperOrigin-RevId: 809691195