[export] Add `fetch_latent_see_json` tool to gRPC Sleuth.
This change introduces a new tool to the `sleuth` utility, allowing users to fetch latent see data from a gRPC server. The `fetch_latent_see_json` tool connects to a specified target, samples latent see data, and outputs it in JSON format to a file or stdout. The `Client` class is updated to include a `LatentSee` stub and a method to call the latent see service.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/812848708](http://cl/812848708) [cl/812696939](http://cl/812696939)

PiperOrigin-RevId: 812848708