[export] [telemetry] Make querying unknown metrics a crash bug
In all cases query code can verify ahead of time whether a metric exists, and so the skip behavior we have is an anti-pattern that hides bugs (including bugs that could lead to telemetry not being populated). Instead, crash on unknown metric queries and force a rapid fix.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/816756414](http://cl/816756414)

PiperOrigin-RevId: 816756414