[export] Introduce UpDownCounter instrument type.
This change adds a new instrument type, `UpDownCounter`, to the gRPC telemetry system. Unlike a standard `Counter`, an `UpDownCounter` can be incremented and decremented. It can be thought of as a UInt Gauge that is stored rather than being queried when the MetricsQuery is run.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/828290610](http://cl/828290610)

PiperOrigin-RevId: 828290610